리팩토링: Calculator, Calculate
- 좀 더 책임에 어울리는 이름으로 클래스명 리팩토링
- Calculator Class -> Runner Class
- Calculate Interface -> Calculator Interface
