--- src/main/java/com/nextstep/calculator/Calculate.java
@@ -1,6 +1,0 @@
-package com.nextstep.calculator;
-
-@FunctionalInterface
-public interface Calculate {
-    Number calculate(Number firstNumber, Number secondNumber);
-}
